

\name{ksl}
\alias{ksl}
\non_function{}
\title{Health and social characteristics}
\description{
 Data from a study measuring health and social characteristics of 
 representative samples of Danish 70 year olds, taken in 1967 and 1984.
}
\format{A data frame with variables of both discrete and continuous
  types.
  \describe{
    \item{FEV}{Forced ejection volume}
    \item{Kol}{Cholesterol}
    \item{Hyp}{Hypertension (no/yes)}
    \item{logBMI}{Logarithm of Body Mass Index}
    \item{Smok}{Smoking (no/yes)}
    \item{Alc}{Alcohol consumption (seldom/frequently)}
    \item{Work}{Working (yes/no)}
    \item{Sex}{male/female}
    \item{Year}{Survey year (1967/1984)}
    }
 }
\references{
%  Morrison, D.F. (1976). Multivariate Statistical Methods. McGraw-Hill, USA.

%  Edwards, D. (1995). Introduction to Graphical Modelling,
%  Springer-Verlag. New York.

  Further information about \bold{deal} can be found at:\cr
  \url{http://www.math.aau.dk/~dethlef/novo/deal}.

  
}
\keyword{datasets}

