%                               -*- Mode: Rd -*- 
% line.Rd --- 
% Author          : Claus Dethlefsen
% Created On      : Sat May 25 23:02:30 2002
% Last Modified By: Claus Dethlefsen
% Last Modified On: Mon May 27 13:27:46 2002
% Update Count    : 2
% Status          : Unknown, Use with caution!
% 

\name{line}
\alias{line}

%- Also NEED an `\alias' for EACH other topic documented here.
\description{Prints a line of symbols}

\title{Prints a line of symbols}

\usage{
line(s="-",n=60)
}

%- maybe also `usage' for other objects documented here.
\arguments{
  \item{s}{Symbol.}
  \item{n}{Number of symbols.}
}

\author{
  Susanne Gammelgaard Bøttcher \email{alma@math.auc.dk}, \cr
  Claus Dethlefsen \email{dethlef@math.auc.dk}.
}

\references{
  Further information about \bold{Deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.
}

%\note{ ~~further notes~~ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\examples{
line()
}

\keyword{models}
