%                               -*- Mode: Rd -*- 
% splash.Rd --- 
% Author          : Claus Dethlefsen
% Created On      : Sat May 25 23:04:32 2002
% Last Modified By: Claus Dethlefsen
% Last Modified On: Sun May 26 11:04:55 2002
% Update Count    : 2
% Status          : Unknown, Use with caution!
% 

\name{splash}
\alias{splash}

%- Also NEED an `\alias' for EACH other topic documented here.

\title{Show a splash screen}

\description{Shows a splash screen first time \bold{Deal} is started.
}

\usage{
splash()
}

%- maybe also `usage' for other objects documented here.

\details{
  Shows a splash-screen and sets the variable nosplash, so that next
  time \bold{Deal} is loaded, the splash-screen is not shown.
}

\author{
  Susanne Gammelgaard Bøttcher \email{alma@math.auc.dk}, \cr
  Claus Dethlefsen \email{dethlef@math.auc.dk}.
}

%\note{ ~~further notes~~ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~
\references{
  Further information about \bold{Deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.
}

\examples{
\dontrun{splash()}
}

\keyword{iplot}
