

\name{ksl}
\alias{ksl}
\title{Health and social characteristics}
\description{
 Data from a study measuring health and social characteristics of 
 representative samples of Danish 70 year olds, taken in 1967 and 1984.
}
\usage{data(ksl)}
\format{A data frame with variables of both discrete and continuous
  types.
  FEV (Forced ejection volume),
  Kol (Cholesterol),
  Hyp (Hypertension (no/yes)),
  BMI (Body Mass Index),
  Smok (Smoking (no/yes)),
  Alc (Alcohol consumption (seldom/frequently)),
  Work (Working (yes/no)), 
  Sex (male/female),
  Year (Survey year (1967/1984)),   
 }
\references{
%  Morrison, D.F. (1976). Multivariate Statistical Methods. McGraw-Hill, USA.

%  Edwards, D. (1995). Introduction to Graphical Modelling,
%  Springer-Verlag. New York.

  Further information about \bold{Deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.

  
}
\keyword{datasets}

