

\name{ksl}
\alias{ksl}
\title{Health and social characteristics}
\description{
 Data from a study measuring health and social characteristics of 
 representative samples of Danish 70 year olds, taken in 1967 and 1984.
}
\usage{data(ksl)}
\format{A dataframe with variables of both discrete and continuous
types. Gender, Year, Smoking, Alcohol, Work, BMI, FEV (lung),
Kolestor, Hyperten}
\references{
  Morrison, D.F. (1976). Multivariate Statistical Methods. McGraw-Hill, USA.

  Edwards, D. (1995). Introduction to Graphical Modelling,
  Springer-Verlag. New York.

  Further information about \bold{Deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.

  
}
\keyword{datasets}

