%                               -*- Mode: Rd -*- 
% line.Rd --- 
% Author          : Claus Dethlefsen
% Created On      : Sat May 25 23:02:30 2002
% Last Modified By: Claus Dethlefsen
% Last Modified On: Wed Jul 23 11:58:46 2003
% Update Count    : 4
% Status          : Unknown, Use with caution!
% 

\name{line}
\alias{line}

%- Also NEED an `\alias' for EACH other topic documented here.
\description{Prints a line of symbols}

\title{Prints a line of symbols}

\usage{
line(s="-",n=60)
}

%- maybe also `usage' for other objects documented here.
\arguments{
  \item{s}{a string giving the symbol.}
  \item{n}{an integer, giving the number of symbols.}
}

\author{
  Susanne Gammelgaard Bøttcher \email{alma@math.auc.dk}, \cr
  Claus Dethlefsen \email{dethlef@math.auc.dk}.
}

\references{
  Further information about \bold{deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.
}

%\note{ ~~further notes~~ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\examples{
line()
}

\keyword{models}
