%                               -*- Mode: Rd -*- 
% rats.Rd --- 
% Author          : Claus Dethlefsen
% Created On      : Sat May 25 23:04:20 2002
% Last Modified By: Claus Dethlefsen
% Last Modified On: Wed Mar 16 13:54:36 2005
% Update Count    : 12
% Status          : Unknown, Use with caution!
% 

\name{rats}
\alias{rats}
\non_function{}
\title{Weightloss of rats}
\description{
  An artificial data set. 24 rats (12 female, 12 male) have been
  randomized to use one of three drugs (products for loosing
  weight). The weightloss for each rat is noted after one and two
  weeks. 
}
\format{A data frame with 4 variables.
  \describe{
    \item{Sex}{a factor with two levels: "M" (male), "F" (female)}
    \item{Drug}{a factor with three levels: "D1", "D2", "D3" (three types)}
    \item{W1}{a numeric: weightloss, week one.}
    \item{W2}{a numeric: weightloss, week 2.}
  }
  }
\references{
  Morrison, D.F. (1976). Multivariate Statistical Methods. McGraw-Hill, USA.

  Edwards, D. (1995). Introduction to Graphical Modelling,
  Springer-Verlag. New York.

  
}
\keyword{datasets}

