%                               -*- Mode: Rd -*- 
% rats.Rd --- 
% Author          : Claus Dethlefsen
% Created On      : Sat May 25 23:04:20 2002
% Last Modified By: Claus Dethlefsen
% Last Modified On: Sat May 25 23:04:21 2002
% Update Count    : 1
% Status          : Unknown, Use with caution!
% 

\name{rats}
\alias{rats}
\title{Weightloss of rats}
\description{
  An artificial dataset. 24 rats (12 female, 12 male) have been
  randomized to use one of three drugs (products for loosing
  weight). The weightloss for each rat is noted after one and two
  weeks. 
}
\usage{data(rats)}
\format{A dataframe with 4 variables. Sex: "M" (male), "F"
  (female). Drug:  "D1", "D2", "D3" (three types). W1 weightloss, week
  one. W2 weightloss, week 2.}
\references{
  Morrison, D.F. (1976). Multivariate Statistical Methods. McGraw-Hill, USA.

  Edwards, D. (1995). Introduction to Graphical Modelling,
  Springer-Verlag. New York.

  Further information about \bold{Deal} can be found at:\cr
  \url{http://www.math.auc.dk/novo/deal}.

  
}
\keyword{datasets}

